library/requirements.txt
asgiref==3.5.2
astroid==2.12.10
async-timeout==4.0.2
backports.zoneinfo==0.2.1
black==22.8.0
click==8.1.3
colorama==0.4.5
Deprecated==1.2.13
dill==0.3.5.1
Django==4.1.2
django-environ==0.9.0
django-redis==5.2.0
djangorestframework==3.14.0
flake8==5.0.4
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pycodestyle==2.9.1
pyflakes==2.5.0
pylint==2.15.3
pylint-plugin-utils==0.7
pyparsing==3.0.9
pytz==2022.4
redis==4.3.4
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.5
typing_extensions==4.4.0
tzdata==2022.4
wrapt==1.14.1
